moved user-directory services to confapi-commons lib including corresponding beans and unit-tests
