+++ src/test/java/de/aservo/atlassian/confluence/confapi/rest/UserDirectoryResourceTest.java
@@ -4,8 +4,8 @@ import com.atlassian.crowd.embedded.api.Directory;
 import com.atlassian.crowd.embedded.api.DirectoryType;
 import com.atlassian.crowd.exception.DirectoryCurrentlySynchronisingException;
 import com.atlassian.crowd.model.directory.DirectoryImpl;
-import de.aservo.atlassian.confluence.confapi.model.UserDirectoryBean;
-import de.aservo.atlassian.confluence.confapi.service.UserDirectoryService;
+import de.aservo.atlassian.confapi.model.UserDirectoryBean;
+import de.aservo.atlassian.confapi.service.UserDirectoryService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
