+++ src/test/java/de/aservo/atlassian/confluence/confapi/suite/TestSuite.java
@@ -5,7 +5,6 @@ import de.aservo.atlassian.confluence.confapi.model.*;
 import de.aservo.atlassian.confluence.confapi.rest.*;
 import de.aservo.atlassian.confluence.confapi.service.ApplicationLinksServiceTest;
 import de.aservo.atlassian.confluence.confapi.service.ExternalGadgetServiceTest;
-import de.aservo.atlassian.confluence.confapi.service.UserDirectoryServiceTest;
 import de.aservo.atlassian.confluence.confapi.service.UserServiceTest;
 import de.aservo.atlassian.confluence.confapi.util.MailProtocolUtilTest;
 import org.junit.runner.RunWith;
@@ -20,7 +19,6 @@ import org.junit.runners.Suite.SuiteClasses;
         ApplicationLinkBeanTest.class,
         LicenseBeanTest.class,
         UserBeanTest.class,
-        UserDirectoryBeanTest.class,
 
         PingResourceTest.class,
         SettingsResourceTest.class,
@@ -34,7 +32,6 @@ import org.junit.runners.Suite.SuiteClasses;
 
         ApplicationLinksServiceTest.class,
         ExternalGadgetServiceTest.class,
-        UserDirectoryServiceTest.class,
         UserServiceTest.class,
 
         AdminOnlyResourceFilterTest.class,
